+++ packages/kobber-icons/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @gyldendal/kobber-components
 
+## 1.0.0
+
+### Major Changes
+
+- Added chevron right and left icons
+
+### Patch Changes
+
+- Updated dependencies
+  - @gyldendal/kobber-base@1.0.0
+
 ## 0.3.87
 
 ### Patch Changes